+++ sdk/identity/Azure.Identity/src/CloudShellManagedIdentitySource.cs
@@ -10,14 +10,13 @@ using Azure.Core.Pipeline;
 
 namespace Azure.Identity
 {
-    internal class CloudShellManagedIdentitySource : IManagedIdentitySource
+    internal class CloudShellManagedIdentitySource : ManagedIdentitySource
     {
-        private readonly HttpPipeline _pipeline;
         private readonly Uri _endpoint;
         private readonly string _clientId;
         private const string MsiEndpointInvalidUriError = "The environment variable MSI_ENDPOINT contains an invalid Uri.";
 
-        public static IManagedIdentitySource TryCreate(HttpPipeline pipeline, string clientId)
+        public static ManagedIdentitySource TryCreate(ManagedIdentityClientOptions options)
         {
             string msiEndpoint = EnvironmentVariables.MsiEndpoint;
 
@@ -37,22 +36,21 @@ namespace Azure.Identity
                 throw new AuthenticationFailedException(MsiEndpointInvalidUriError, ex);
             }
 
-            return new CloudShellManagedIdentitySource(pipeline, endpointUri, clientId);
+            return new CloudShellManagedIdentitySource(options.Pipeline, endpointUri, options.ClientId);
         }
 
-        private CloudShellManagedIdentitySource(HttpPipeline pipeline, Uri endpoint, string clientId)
+        private CloudShellManagedIdentitySource(CredentialPipeline pipeline, Uri endpoint, string clientId) : base(pipeline)
         {
-            _pipeline = pipeline;
             _endpoint = endpoint;
             _clientId = clientId;
         }
 
-        public Request CreateRequest(string[] scopes)
+        protected override Request CreateRequest(string[] scopes)
         {
             // covert the scopes to a resource string
             string resource = ScopeUtilities.ScopesToResource(scopes);
 
-            Request request = _pipeline.CreateRequest();
+            Request request = Pipeline.HttpPipeline.CreateRequest();
 
             request.Method = RequestMethod.Post;
 
@@ -73,20 +71,5 @@ namespace Azure.Identity
             request.Content = RequestContent.Create(content);
             return request;
         }
-
-        public AccessToken GetAccessTokenFromJson(in JsonElement jsonAccessToken, in JsonElement jsonExpiresOn)
-        {
-            // the seconds from epoch may be returned as a Json number or a Json string which is a number
-            // depending on the environment.  If neither of these are the case we throw an AuthException.
-            if (jsonExpiresOn.ValueKind == JsonValueKind.Number && jsonExpiresOn.TryGetInt64(out long expiresOnSec) ||
-                jsonExpiresOn.ValueKind == JsonValueKind.String && long.TryParse(jsonExpiresOn.GetString(), out expiresOnSec))
-            {
-                return new AccessToken(jsonAccessToken.GetString(), DateTimeOffset.FromUnixTimeSeconds(expiresOnSec));
-            }
-
-            throw new AuthenticationFailedException(ManagedIdentityClient.AuthenticationResponseInvalidFormatError);
-        }
-
-        public ValueTask HandleFailedRequestAsync(Response response, ClientDiagnostics diagnostics, bool async) => new ValueTask();
     }
 }