+++ sdk/identity/Azure.Identity/src/ImdsManagedIdentitySource.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Net;
 using System.Net.Sockets;
+using System.Runtime.CompilerServices;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
@@ -12,7 +13,7 @@ using Azure.Core.Pipeline;
 
 namespace Azure.Identity
 {
-    internal class ImdsManagedIdentitySource : IManagedIdentitySource
+    internal class ImdsManagedIdentitySource : ManagedIdentitySource
     {
         // IMDS constants. Docs for IMDS are available here https://docs.microsoft.com/en-us/azure/active-directory/managed-identities-azure-resources/how-to-use-vm-token#get-a-token-using-http
         private static readonly Uri s_imdsEndpoint = new Uri("http://169.254.169.254/metadata/identity/oauth2/token");
@@ -23,12 +24,11 @@ namespace Azure.Identity
 
         internal const string IdentityUnavailableError = "ManagedIdentityCredential authentication unavailable. The requested identity has not been assigned to this resource.";
 
-        private readonly HttpPipeline _pipeline;
         private readonly string _clientId;
 
         private string _identityUnavailableErrorMessage;
 
-        public static async ValueTask<IManagedIdentitySource> TryCreateAsync(HttpPipeline pipeline, string clientId, bool async, CancellationToken cancellationToken)
+        public static async ValueTask<ManagedIdentitySource> TryCreateAsync(ManagedIdentityClientOptions options, bool async, CancellationToken cancellationToken)
         {
             AzureIdentityEventSource.Singleton.ProbeImdsEndpoint(s_imdsEndpoint);
 
@@ -68,16 +68,15 @@ namespace Azure.Identity
                 AzureIdentityEventSource.Singleton.ImdsEndpointUnavailable(s_imdsEndpoint);
             }
 
-            return available ? new ImdsManagedIdentitySource(pipeline, clientId) : default;
+            return available ? new ImdsManagedIdentitySource(options.Pipeline, options.ClientId) : default;
         }
 
-        internal ImdsManagedIdentitySource(HttpPipeline pipeline, string clientId)
+        internal ImdsManagedIdentitySource(CredentialPipeline pipeline, string clientId) : base(pipeline)
         {
-            _pipeline = pipeline;
             _clientId = clientId;
         }
 
-        public Request CreateRequest(string[] scopes)
+        protected override Request CreateRequest(string[] scopes)
         {
             if (_identityUnavailableErrorMessage != default)
             {
@@ -87,7 +86,7 @@ namespace Azure.Identity
             // covert the scopes to a resource string
             string resource = ScopeUtilities.ScopesToResource(scopes);
 
-            Request request = _pipeline.CreateRequest();
+            Request request = Pipeline.HttpPipeline.CreateRequest();
             request.Method = RequestMethod.Get;
             request.Headers.Add("Metadata", "true");
             request.Uri.Reset(s_imdsEndpoint);
@@ -103,30 +102,19 @@ namespace Azure.Identity
             return request;
         }
 
-        public AccessToken GetAccessTokenFromJson(in JsonElement jsonAccessToken, in JsonElement jsonExpiresOn)
-        {
-            // the seconds from epoch may be returned as a Json number or a Json string which is a number
-            // depending on the environment.  If neither of these are the case we throw an AuthException.
-            if (jsonExpiresOn.ValueKind == JsonValueKind.Number && jsonExpiresOn.TryGetInt64(out long expiresOnSec) ||
-                jsonExpiresOn.ValueKind == JsonValueKind.String && long.TryParse(jsonExpiresOn.GetString(), out expiresOnSec))
-            {
-                return new AccessToken(jsonAccessToken.GetString(), DateTimeOffset.FromUnixTimeSeconds(expiresOnSec));
-            }
-
-            throw new AuthenticationFailedException(ManagedIdentityClient.AuthenticationResponseInvalidFormatError);
-        }
-
-        public async ValueTask HandleFailedRequestAsync(Response response, ClientDiagnostics diagnostics, bool async)
+        protected override async ValueTask<AccessToken> HandleResponseAsync(bool async, TokenRequestContext context, Response response, CancellationToken cancellationToken)
         {
             if (response.Status == 400)
             {
-                string message = _identityUnavailableErrorMessage ?? await diagnostics
+                string message = _identityUnavailableErrorMessage ?? await Pipeline.Diagnostics
                     .CreateRequestFailedMessageAsync(response, IdentityUnavailableError, null, null, async)
                     .ConfigureAwait(false);
 
                 Interlocked.CompareExchange(ref _identityUnavailableErrorMessage, message, null);
                 throw new CredentialUnavailableException(message);
             }
+
+            return await base.HandleResponseAsync(async, context, response, cancellationToken).ConfigureAwait(false);
         }
     }
 }