+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.3.0-beta.3 (Unreleased)
 
+### New Features
+- Added support for Service Fabric managed identity authentication to `ManagedIdentityCredential`.
+- Added support for Azure Arc managed identity authentication to `ManagedIdentityCredential`.
+
 ### Fixes and improvements
 - Fix race condition in `ProcessRunner` causing `VisualStudioCredential` and `AzureCliCredential` to fail intermittently ([#16211](https://github.com/Azure/azure-sdk-for-net/issues/16211))
 - Fix `VisualStudioCodeCredential` to raise `CredentialUnavailableException` when reads from libsecret fail ([#16795](https://github.com/Azure/azure-sdk-for-net/issues/16795))